+++ CHANGELOG.md
@@ -4,4 +4,5 @@
 ## 0.1.0 (unreleased)
 * Improvements to web interface
 * Switch to pydantic
-
+* Server protocol v1 (the original protocol is still supported)
+* Basic testing for Python clients